namespaces' scope is now correctly represented
Namespaces are not attached to the document anymore, insted they are
now bound to the node which defines the new namespace.
All descendant will have a reference to all namespaces valid in their
scope. All namespaces valid in a certain scope are stored in
node->knownNamespaces grouped using an XmlNamespaceSet linkedlist